fix: fallback width on terminals with no reported width
Some terminals (like the JetBrains output panel) do not support
the Linux technique for polling the width. It returns 0 as a result.
Since 0 makes no sense as a real-world terminal width, we take this
return value as a signal the width could not be polled. If this happens,
we use the fallback width of 32.